+++ go/worker/plugins/ray/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//go/worker/activities/ray:go_default_library",
+        "//go/worker/plugins/utils:go_default_library",
         "//proto/api:go_default_library",
         "//proto/api/v2:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//cadstar:go_default_library",
@@ -32,6 +33,7 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":go_default_library"],
     deps = [
+        "@com_github_cadence_workflow_starlark_worker//cadstar:go_default_library",
         "@com_github_stretchr_testify//suite:go_default_library",
     ],
 )
