+++ go/worker/plugins/ray/starlark_module.go
@@ -6,8 +6,6 @@ import (
 
 	"github.com/cadence-workflow/starlark-worker/cadstar"
 	"github.com/cadence-workflow/starlark-worker/ext"
-	"github.com/cadence-workflow/starlark-worker/star"
-	jsoniter "github.com/json-iterator/go"
 	"go.starlark.net/starlark"
 	"go.uber.org/cadence"
 	"go.uber.org/cadence/workflow"
@@ -16,6 +14,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/michelangelo-ai/michelangelo/go/worker/activities/ray"
+	"github.com/michelangelo-ai/michelangelo/go/worker/plugins/utils"
 	apipb "github.com/michelangelo-ai/michelangelo/proto/api"
 	v2pb "github.com/michelangelo-ai/michelangelo/proto/api/v2"
 )
@@ -28,44 +27,6 @@ var poll int64 = 10
 
 type module struct {
 	attributes map[string]starlark.Value
-}
-
-const CadenceLongTimeout = time.Hour * 24 * 365 * 10 // 10 years, practically - no timeout
-
-var CadenceDefaultNonRetriableErrorReasons = []string{
-	"cadenceInternal:Panic",                  // panics
-	"cadenceInternal:Generic",                // cadence converter errors (similar to invalid-argument)
-	"400",                                    // bad-request https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/400
-	"401",                                    // unauthorized
-	"403",                                    // forbidden
-	"404",                                    // not-found
-	"405",                                    // method-not-allowed
-	"502",                                    // bad-gateway
-	yarpcerrors.CodeCancelled.String(),       // client error
-	yarpcerrors.CodeNotFound.String(),        // client error
-	yarpcerrors.CodeAlreadyExists.String(),   // client error
-	yarpcerrors.CodeInvalidArgument.String(), // client error
-	yarpcerrors.CodeUnauthenticated.String(), // client error
-	yarpcerrors.CodePermissionDenied.String(), // client error
-	yarpcerrors.CodeUnimplemented.String(),    // client error
-	yarpcerrors.CodeDataLoss.String(),         // server error; unrecoverable data corruption
-	yarpcerrors.CodeInternal.String(),         // server error; serious error, like panic
-}
-
-var CadenceDefaultRetryPolicy = workflow.RetryPolicy{
-	InitialInterval:          time.Second * 15,
-	BackoffCoefficient:       1,
-	ExpirationInterval:       time.Minute * 5,
-	NonRetriableErrorReasons: CadenceDefaultNonRetriableErrorReasons,
-	MaximumAttempts:          1,
-}
-
-var CadenceDefaultSensorRetryPolicy = workflow.RetryPolicy{
-	InitialInterval:          time.Second * 10,
-	BackoffCoefficient:       1,
-	ExpirationInterval:       CadenceLongTimeout,
-	NonRetriableErrorReasons: CadenceDefaultNonRetriableErrorReasons,
-	MaximumAttempts:          100,
 }
 
 func newModule() starlark.Value {
@@ -85,20 +46,6 @@ func (r *module) Truth() starlark.Bool                  { return true }
 func (r *module) Hash() (uint32, error)                 { return 0, fmt.Errorf("no-hash") }
 func (r *module) Attr(n string) (starlark.Value, error) { return r.attributes[n], nil }
 func (r *module) AttrNames() []string                   { return ext.SortedKeys(r.attributes) }
-func AsStar(source any, out any) error {
-	b, err := jsoniter.Marshal(source)
-	if err != nil {
-		return err
-	}
-	return star.Decode(b, out)
-}
-func AsGo(source starlark.Value, out any) error {
-	b, err := star.Encode(source)
-	if err != nil {
-		return err
-	}
-	return jsoniter.Unmarshal(b, out)
-}
 
 func (r *module) createCluster(t *starlark.Thread, _ *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
 	ctx := cadstar.GetContext(t)
@@ -111,7 +58,7 @@ func (r *module) createCluster(t *starlark.Thread, _ *starlark.Builtin, args sta
 	}
 
 	var cluster v2pb.RayCluster
-	if err := AsGo(spec, &cluster); err != nil {
+	if err := utils.AsGo(spec, &cluster); err != nil {
 		logger.Error("builtin-error", ext.ZapError(err)...)
 		return nil, err
 	}
@@ -124,7 +71,7 @@ func (r *module) createCluster(t *starlark.Thread, _ *starlark.Builtin, args sta
 
 	cluster = *response.RayCluster
 
-	srp := CadenceDefaultSensorRetryPolicy
+	srp := utils.CadenceDefaultSensorRetryPolicy
 	srp.ExpirationInterval = time.Second * time.Duration(timeout)
 	srp.InitialInterval = time.Second * time.Duration(poll)
 	sensorCtx := workflow.WithRetryPolicy(ctx, srp)
@@ -176,7 +123,7 @@ func (r *module) createCluster(t *starlark.Thread, _ *starlark.Builtin, args sta
 
 	sensorCluster := sensorResponse.RayCluster
 	var res starlark.Value
-	if err := AsStar(sensorCluster, &res); err != nil {
+	if err := utils.AsStar(sensorCluster, &res); err != nil {
 		logger.Error("builtin-error", ext.ZapError(err)...)
 		return nil, err
 	}
@@ -227,7 +174,7 @@ func (r *module) createJob(t *starlark.Thread, _ *starlark.Builtin, args starlar
 	rayJob = *createRes.RayJob
 
 	var sensorRes ray.SensorRayJobResponse
-	srp := CadenceDefaultSensorRetryPolicy
+	srp := utils.CadenceDefaultSensorRetryPolicy
 	srp.ExpirationInterval = time.Second * time.Duration(timeout)
 	srp.InitialInterval = time.Second * time.Duration(poll)
 	sensorCtx := workflow.WithRetryPolicy(ctx, srp)
@@ -242,7 +189,7 @@ func (r *module) createJob(t *starlark.Thread, _ *starlark.Builtin, args starlar
 
 	job := sensorRes.RayJob
 	var res starlark.Value
-	if err := AsStar(job, &res); err != nil {
+	if err := utils.AsStar(job, &res); err != nil {
 		logger.Error("builtin-error", ext.ZapError(err)...)
 		return nil, err
 	}
@@ -269,7 +216,7 @@ func (r *module) terminateCluster(t *starlark.Thread, _ *starlark.Builtin, args 
 	}
 
 	var res v2pb.UpdateRayClusterResponse
-	srp := CadenceDefaultSensorRetryPolicy
+	srp := utils.CadenceDefaultSensorRetryPolicy
 	srp.ExpirationInterval = time.Second * time.Duration(timeout)
 	srp.InitialInterval = time.Second * time.Duration(poll)
 	sensorCtx := workflow.WithRetryPolicy(ctx, srp)
