+++ go/worker/plugins/utils/BUILD.bazel
@@ -1,0 +1,15 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["utils.go"],
+    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/plugins/utils",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@com_github_cadence_workflow_starlark_worker//star:go_default_library",
+        "@com_github_json_iterator_go//:go_default_library",
+        "@net_starlark_go//starlark:go_default_library",
+        "@org_uber_go_cadence//workflow:go_default_library",
+        "@org_uber_go_yarpc//yarpcerrors:go_default_library",
+    ],
+)
